refactor: state 폴더 생성
