+++ src/main/java/blackjack/model/state/Hit.java
@@ -1,5 +1,7 @@
-package blackjack.model.card;
+package blackjack.model.state;
 
+import blackjack.model.card.Card;
+import blackjack.model.card.Hand;
 import blackjack.model.gameresult.GameResult;
 
 public class Hit implements HandState {
